removed names from about page
